feat: 팀 페이지 마크업 - 홈 페이지 마크업 > 컴포넌트로 나누기 - 캐러셀 화살표 버튼 컴포넌트 추가 - 아이콘을 포함하는 라운드 버튼 컴포넌트 추가
