+++ src/app/(auth)/signin/page.tsx
@@ -1,5 +1,5 @@
 'use client';
-import { useAuthStore } from '@/stores/useAuthStore';
+import { useAuthStore } from '@stores/useAuthStore';
 import React from 'react';
 import Image from 'next/image';
 import { Logo } from '@assets/icons';
@@ -7,6 +7,7 @@ import Input from '@components/forms/inputGroups/Input';
 import { LoginButton } from '@components/elements/buttons/LoginButton';
 import useStore from '@hooks/useStore';
 import { useRouter } from 'next/navigation';
+import Link from 'next/link';
 
 function Login() {
 	// MARK :: zustand persist 적용 시, Error 발생
@@ -23,9 +24,9 @@ function Login() {
 	return (
 		<div className="flex w-screen h-screen flex-col items-center justify-center">
 			<div className="rem:w-[590px]">
-				<div className="logo flex items-center justify-center rem:pb-[64px] ">
+				<Link href="/home" className="logo flex items-center justify-center rem:pb-[64px] ">
 					<Image src={Logo} alt="logo" className="rem:w-[170px] rem:h-[35px]" />
-				</div>
+				</Link>
 
 				<div className="grid rem:gap-[20px] rem:pb-[30px]">
 					<Input type="text" placeholder="아이디" value={store.id} onChange={(e) => store.setId(e.target.value)} />
