+++ src/app/globals.css
@@ -31,7 +31,7 @@ html {
 	}
 }
 
-.text-ellipsis {
+.text-ellipsis-line-1 {
 	overflow: hidden;
 	white-space: nowrap;
 	text-overflow: ellipsis;
