+++ tsconfig.json
@@ -26,6 +26,7 @@
 			"@hooks/*": ["./src/hooks/*"],
 			"@models/*": ["./src/models/*"],
 			"@utils/*": ["./src/utils/*"],
+			"@stores/*": ["./src/stores/*"],
 			"@/*": ["./src/*"]
 		}
 	},
